requirements.txt
Django==1.9.5
defusedxml==0.4.1
django-allauth==0.24.1
django-rest-auth==0.7.0
djangorestframework==3.3.3
gunicorn==19.4.5
oauthlib==1.0.3
psycopg2>=2.6.1
python3-openid==3.0.10
requests==2.9.1
requests-oauthlib==0.6.1
six==1.9.0
